+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+### 0.0.9
+
+* Fixed OkpoValidator
+
+### 0.0.8
+
+* Validate the Crimea region numbers
+* Speeding up validators
+
 ### 0.0.7
 
 * BikValidator